+++ lib/puppet/provider/oneview_hypervisor_cluster_profile/c7000.rb
@@ -25,7 +25,7 @@ Puppet::Type.type(:oneview_hypervisor_cluster_profile).provide :c7000, parent: P
   mk_resource_methods
 
   def self.api_version
-    800
+    1800
   end
 
   def self.resource_name
@@ -33,6 +33,7 @@ Puppet::Type.type(:oneview_hypervisor_cluster_profile).provide :c7000, parent: P
   end
 
   def destroy
+    Puppet.info "Data Value: #{@data}"
     force = @data.delete('force') || false
     soft_delete = @data.delete('soft_delete') || false
     hypervisor_cluster_profile = get_single_resource_instance