+++ CHANGELOG.md
@@ -9,9 +9,11 @@
 - Ethernet Network
 - FC Network
 - FCOE Network
+- Hypervisor Cluster Profile
 - Hypervisor Manager
 - Interconnect
 - Logical Enclosure
+- Logical Interconnect
 - Logical Interconnect Group
 - Network Set
 - Server Certificate